+++ smallworld/SmallworldLogicNumba.py
@@ -418,7 +418,7 @@ class Board():
 			else:
 				# https://stackoverflow.com/questions/3062746/special-simple-random-number-generator
 				# m=6, c=5, a=1980+1
-				rnd_value = (1981 * (random_seed+self.invisible_deck[5]) + 5) % 6
+				rnd_value = (1981 * (random_seed+np.int64(self.invisible_deck[5])) + 5) % 6
 				dice = DICE_VALUES[rnd_value]
 			self.invisible_deck[5] += 1
 			if nb_ppl_of_player + dice < minimum_ppl_for_attack:
@@ -1186,7 +1186,7 @@ class Board():
 			else:
 				# https://stackoverflow.com/questions/3062746/special-simple-random-number-generator
 				# m=6, c=5, a=1980+1
-				rnd_value = (1981 * (random_seed+self.invisible_deck[5]) + 5) % 6
+				rnd_value = (1981 * (random_seed+np.int64(self.invisible_deck[5])) + 5) % 6
 				dice = DICE_VALUES[rnd_value]
 			self.invisible_deck[5] += 1
 			current_ppl[4] = dice + 2**6
@@ -1365,9 +1365,9 @@ class Board():
 			else:
 				# https://stackoverflow.com/questions/3062746/special-simple-random-number-generator
 				# m=avail_people_id.size, c=0, a=2*3*5*7*9*11*13*17+1
-				rnd_value = (4594591 * (random_seed+self.invisible_deck[6])) % avail_people_id.size
+				rnd_value = (4594591 * (random_seed+np.int64(self.invisible_deck[6]))) % avail_people_id.size
 				chosen_ppl = avail_people_id[rnd_value]
-				rnd_value = (4594591 * (random_seed+self.invisible_deck[6])) % avail_power_id.size
+				rnd_value = (4594591 * (random_seed+np.int64(self.invisible_deck[6]))) % avail_power_id.size
 				chosen_power = avail_power_id[rnd_value]
 			self.invisible_deck[6] += 1
 			nb_of_ppl = initial_nb_people[chosen_ppl] + initial_nb_power[chosen_power]
@@ -1408,9 +1408,9 @@ class Board():
 					else:
 						# https://stackoverflow.com/questions/3062746/special-simple-random-number-generator
 						# m=avail_people_id.size, c=0, a=2*3*5*7*9*11*13*17+1
-						rnd_value = (4594591 * (random_seed+self.invisible_deck[6])) % avail_people_id.size
+						rnd_value = (4594591 * (random_seed+np.int64(self.invisible_deck[6]))) % avail_people_id.size
 						chosen_ppl = avail_people_id[rnd_value]
-						rnd_value = (4594591 * (random_seed+self.invisible_deck[6])) % avail_power_id.size
+						rnd_value = (4594591 * (random_seed+np.int64(self.invisible_deck[6]))) % avail_power_id.size
 						chosen_power = avail_power_id[rnd_value]
 					self.invisible_deck[6] += 1
 					nb_of_ppl = initial_nb_people[chosen_ppl] + initial_nb_power[chosen_power]
